pyproject.toml
[tool.poetry]
name = "market-info-api"
version = "0.1.0"
description = "Django REST API for fetching and storing website information"
authors = ["Alexander Yuvcehnko <alex.yuvchenko@gmail.com>"]
readme = "README.md"
packages = [{include = "apps"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.7"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
validators = "^0.34.0"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.10.0"
pytest-cov = "^6.0.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "apps.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -v --cov=apps --cov-report=term-missing"
markers = [
    "django_db: Mark a test as requiring the database",
]
testpaths = ["tests"]
